Homepage is just about done. Created profile page.
Homepage is good enough to be passable... Could use some more modules on
the right side though, but I'm going to start making some progress on the
profile page first.

Started the initial effort of creating the top module of the profile page
with the person's name, photo, and basic summary.